Initial commit of ScratchPad plugin.
The plugin makes it easy to store links and quotes from web sites for you to quote in your blog posts. Drag the bookmarklet to your toolbar. When you're on a page you want to refer to in a post, highlight text to quote and click the bookmarklet. A form will appear in the top right of the page with information about the page. You can save the information in the default ScratchPad or create a named ScratchPad. When you go to the publish page you will have a ScratchPad media silo that contains the links and data you have collected, organised by the name of the ScratchPad. Each ScratchPad can have a template, so it can be inserted in a post differently (though the management interface for the templates is not yet written).

An example usage might be that you're on a TechCrunch page about Twitter downtime and there's a particularly mean quote. You highlight the quote and save it under a ScratchPad named 'Twitter woes'. You then surf around a bit more, finding other people with supporting or opposing views, and saving links, with or without quotes to your 'Twitter woes' ScratchPad. You then go to your publish page, open the ScratchPads media silo and select the 'Twitter woes' directory. You can insert your links, quotes, title of the pages, with on click.

Another use is to store links for sites you reference regularly, for example under a named ScratchPads called 'reference'. This might have simple links such as <a href="http://habariproject.org" title="Habari Project">Habari Project</a>.

To do: management interface for adding and removing ScratchPads or setting templates, make the bookmarklet look good, enable posting when the user isn't logged in (ie allow them to log in), fine tune silo display.


git-svn-id: http://svn.habariproject.org/habari-extras@632 f4a90942-1786-4aa1-b092-57c31935ddc3