plugin:staticcache NOTE: This update changes the basic structure of the cache store. Please delete the cache or you will probably get errors. There should a button to invalidate staticcaches cache, but there isn't yet. If you want, please add one :).
Changing cache structure to include the caching of 
headers to help prevent improper caching on redirects. Unfortunately we 
still cannot grab the status header, so it will send 200, or 302 for 
redirects. Also adding 404 detection, via the "matched rule", to ensure 
that 404 response will not be cached and served as 200.

There is also now some basic hits/misses stats to show the percentage of 
requests that are served cached pages (hits), and percentage that are 
not (misses).


git-svn-id: http://svn.habariproject.org/habari-extras@922 f4a90942-1786-4aa1-b092-57c31935ddc3