Oops. Here are the changes I announced in 3540. Tired me uploaded the local version that did not include the changes.
Additionally, I removed the second paragraph in the authorization confirmation message as the deauthorization url had an error.

Here is the original text from 3540:

Fixed the Javascript for inserting a url around the image. The preview-Javascript still has to be fixed.

Added support for various sizes. This can be extended as Picasa supports any sizes up to the original size (no steps required as for Flickr). Changed the inserted code so it now links to the original size and the image only shows the size from the settings (as the Flickr silo does).

Added language support (might be incomplete).

git-svn-id: http://svn.habariproject.org/habari-extras@3541 f4a90942-1786-4aa1-b092-57c31935ddc3